Firewall, antiVirus Bug Fix, Typos.
Few functions added:
1. updateFirewallDirection()
2. GetEnemyType()
3. neighborExist90()
4. neighborExist360()

+++ src/player.cpp
@@ -83,7 +83,7 @@ void player::update() {
 }
 
 void player::draw() {
-    DrawTextureRec(texture, Rectangle{(float) age * 32, 0, 32, 32}, Vector2{(float) posX * 32, (float) posY * 32},
+    DrawTextureRec(texture, Rectangle{age * 32, 0, 32, 32}, Vector2{(float) posX * 32, (float) posY * 32},
                    WHITE);
 
 }
